+++ projects/doc/src/app/components-examples/forms/inputs/input-types-example/input-types-example.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'gd-input-types-example',
+  templateUrl: './input-types-example.component.html',
+  styleUrls: ['./input-types-example.component.scss']
+})
+export class InputTypesExampleComponent {
+
+}