+++ projects/doc/src/app/components-examples/forms/inputs/input-validation-example/input-validation-example.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { InputValidationExampleComponent } from './input-validation-example.component';
+
+describe('InputValidationExampleComponent', () => {
+  let component: InputValidationExampleComponent;
+  let fixture: ComponentFixture<InputValidationExampleComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [InputValidationExampleComponent]
+    });
+    fixture = TestBed.createComponent(InputValidationExampleComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});